chore(temporal): decouple health check framework from dagster
The HealthCheck framework is logically owned by Temporal — every active
check is registered as a `HealthCheck` subclass under
`posthog/temporal/health_checks/` and runs as a Temporal workflow. This
removes the last cross-imports from `posthog.dags.common.health.*` so
the framework is self-contained.

- Inline `HealthCheckResult` and `BatchResult` directly into
  `posthog/temporal/health_checks/models.py` (dropping the dagster-aware
  `BatchDetectFn` and `HealthCheckDefinition` siblings).
- Move `_upsert_issues` / `_resolve_stale_issues` to a new
  `posthog/temporal/health_checks/db.py`.
- Move `push_health_check_metrics` to a new
  `posthog/temporal/health_checks/observability.py`.
- Delete `posthog/dags/common/health/` (entire folder) and its test
  module — the live checks are covered by tests under
  `products/*/backend/temporal/health_checks/`.
- Delete the redundant Dagster jobs
  `products/web_analytics/dags/no_live_events.py` and
  `products/web_analytics/dags/no_pageleave_events.py`; the Temporal
  equivalents are already registered in `HEALTH_CHECK_MODULES`.
- Drop their entries from `posthog/dags/locations/shared.py`.

Generated-By: PostHog Code
Task-Id: 94f67a84-4ad4-457f-88be-91d5ed97a085
