--- posthog/dags/common/health/framework.py
@@ -1,220 +1,0 @@
-from collections.abc import Mapping
-from types import MappingProxyType
-
-import dagster
-
-from posthog.dags.common import JobOwners, check_for_concurrent_runs
-from posthog.dags.common.health.detectors import HealthDetector, resolve_execution_policy
-from posthog.dags.common.health.observability import push_health_check_metrics
-from posthog.dags.common.health.processing import _process_batch_detection
-from posthog.dags.common.health.types import BatchResult, HealthCheckDefinition, HealthCheckThresholdExceeded
-from posthog.dags.common.ops import get_all_team_ids_op
-
-_REGISTERED_KINDS: dict[str, str] = {}
-
-_default_retry_policy: dagster.RetryPolicy = dagster.RetryPolicy(
-    max_retries=2,
-    delay=30,
-    backoff=dagster.Backoff.EXPONENTIAL,
-    jitter=dagster.Jitter.PLUS_MINUS,
-)
-
-
-def get_registered_kinds() -> Mapping[str, str]:
-    return MappingProxyType(_REGISTERED_KINDS)
-
-
-def _reset_registry() -> None:
-    """Clear the kind registry. For test use only."""
-    _REGISTERED_KINDS.clear()
-
-
-def _to_batch_dagster_metadata(result: BatchResult, teams_per_second: float) -> dict[str, dagster.MetadataValue]:
-    return {
-        "detect_duration_seconds": dagster.MetadataValue.float(result.detect_duration),
-        "db_write_duration_seconds": dagster.MetadataValue.float(result.db_write_duration),
-        "resolve_duration_seconds": dagster.MetadataValue.float(result.resolve_duration),
-        "total_duration_seconds": dagster.MetadataValue.float(result.total_duration),
-        "teams_per_second": dagster.MetadataValue.float(teams_per_second),
-        "batch_size": dagster.MetadataValue.int(result.batch_size),
-        "issues_upserted": dagster.MetadataValue.int(result.issues_upserted),
-        "issues_resolved": dagster.MetadataValue.int(result.issues_resolved),
-        "teams_with_issues": dagster.MetadataValue.int(result.teams_with_issues),
-        "teams_healthy": dagster.MetadataValue.int(result.teams_healthy),
-        "teams_failed": dagster.MetadataValue.int(result.teams_failed),
-        "teams_skipped": dagster.MetadataValue.int(result.teams_skipped),
-    }
-
-
-def _to_aggregate_dagster_metadata(totals: BatchResult, teams_per_second: float) -> dict[str, dagster.MetadataValue]:
-    return {
-        "total_teams": dagster.MetadataValue.int(totals.batch_size),
-        "total_issues_upserted": dagster.MetadataValue.int(totals.issues_upserted),
-        "total_issues_resolved": dagster.MetadataValue.int(totals.issues_resolved),
-        "total_teams_with_issues": dagster.MetadataValue.int(totals.teams_with_issues),
-        "total_teams_healthy": dagster.MetadataValue.int(totals.teams_healthy),
-        "total_teams_failed": dagster.MetadataValue.int(totals.teams_failed),
-        "total_teams_skipped": dagster.MetadataValue.int(totals.teams_skipped),
-        "total_duration_seconds": dagster.MetadataValue.float(float(totals.total_duration)),
-        "teams_per_second": dagster.MetadataValue.float(float(teams_per_second)),
-    }
-
-
-def _create_check_batch_op(name: str, kind: str, detector: HealthDetector) -> dagster.OpDefinition:
-    @dagster.op(name=f"{name}_check_batch", retry_policy=_default_retry_policy)
-    def check_batch_op(context: dagster.OpExecutionContext, team_ids: list[int]) -> BatchResult:
-        ops = context.run_config.get("ops", {})
-        dry_run = False
-        if isinstance(ops, dict):
-            team_ids_op = ops.get("get_all_team_ids_op", {})
-            if isinstance(team_ids_op, dict):
-                config = team_ids_op.get("config", {})
-                if isinstance(config, dict):
-                    dry_run = bool(config.get("dry_run", False))
-        result = _process_batch_detection(team_ids, kind, detector.detect_fn, context, dry_run=dry_run)
-
-        context.add_output_metadata(_to_batch_dagster_metadata(result, result.teams_per_second))
-
-        return result
-
-    return check_batch_op
-
-
-def _aggregate_results(
-    results: list[BatchResult],
-    kind: str,
-    not_processed_threshold: float,
-    context: dagster.OpExecutionContext,
-) -> None:
-    totals = BatchResult()
-    for r in results:
-        totals += r
-
-    context.log.info(
-        f"Health check '{kind}' completed: {totals.batch_size:,} teams in {totals.total_duration:.0f}s "
-        f"({totals.teams_per_second:.0f} teams/sec)\n"
-        f"  Issues: {totals.issues_upserted:,} upserted, {totals.issues_resolved:,} resolved\n"
-        f"  Teams: {totals.teams_with_issues:,} with issues, {totals.teams_healthy:,} healthy, "
-        f"{totals.teams_skipped:,} skipped, {totals.teams_failed:,} failed"
-    )
-
-    context.add_output_metadata(_to_aggregate_dagster_metadata(totals, totals.teams_per_second))
-
-    not_processed = totals.teams_skipped + totals.teams_failed
-    threshold_exceeded = totals.batch_size > 0 and totals.not_processed_rate > not_processed_threshold
-
-    push_health_check_metrics(kind, totals, success=not threshold_exceeded)
-
-    if threshold_exceeded:
-        raise HealthCheckThresholdExceeded(
-            f"Health check '{kind}': {not_processed:,}/{totals.batch_size:,} teams not processed "
-            f"({totals.not_processed_rate:.1%}), exceeds threshold {not_processed_threshold:.1%} "
-            f"(skipped={totals.teams_skipped:,}, failed={totals.teams_failed:,})"
-        )
-
-
-def _create_aggregate_op(
-    name: str,
-    kind: str,
-    not_processed_threshold: float,
-) -> dagster.OpDefinition:
-    @dagster.op(name=f"{name}_aggregate")
-    def aggregate_op(context: dagster.OpExecutionContext, results: list[BatchResult]) -> None:
-        _aggregate_results(results, kind, not_processed_threshold, context)
-
-    return aggregate_op
-
-
-def create_health_check(
-    name: str,
-    kind: str,
-    detector: HealthDetector,
-    owner: JobOwners,
-    *,
-    team_ids: list[int] | None = None,
-    schedule: str | None = None,
-    batch_size: int | None = None,
-    max_concurrent: int | None = None,
-    rollout_percentage: float | None = None,
-    not_processed_threshold: float = 0.1,
-    dry_run: bool = False,
-) -> HealthCheckDefinition:
-    """Create a complete health check pipeline as a Dagster job.
-
-    Args:
-        name: Unique name for the health check (used in Dagster op/job names).
-        kind: The health issue kind string written to the health_issues table.
-        detector: Detector definition used to detect health issues.
-        owner: The team that owns this health check for Slack alert routing.
-        team_ids: If provided, only process these team IDs. If None, processes all teams.
-        schedule: Optional cron expression for scheduling.
-        batch_size: Optional override for detector default batch size.
-        max_concurrent: Optional override for detector default concurrency.
-        rollout_percentage: Optional deterministic team rollout fraction (0–1, e.g. 0.01 = 1%).
-        not_processed_threshold: Fail if this fraction of teams are skipped or failed.
-        dry_run: If True, run detection but skip DB writes (upsert/resolve).
-
-    Returns:
-        HealthCheckDefinition
-    """
-    existing = _REGISTERED_KINDS.get(kind)
-    if existing is not None and existing != name:
-        raise ValueError(f"Health check kind '{kind}' already registered by '{existing}'")
-    _REGISTERED_KINDS[kind] = name
-
-    policy = resolve_execution_policy(
-        detector,
-        batch_size=batch_size,
-        max_concurrent=max_concurrent,
-    )
-
-    batch_op = _create_check_batch_op(name, kind, detector)
-    aggregate_op = _create_aggregate_op(name, kind, not_processed_threshold)
-
-    op_config: dict = {"batch_size": policy.batch_size}
-    if team_ids is not None:
-        op_config["team_ids"] = team_ids
-    if rollout_percentage is not None:
-        op_config["rollout_percentage"] = rollout_percentage
-
-    op_config["dry_run"] = dry_run
-
-    job_config = {
-        "ops": {
-            "get_all_team_ids_op": {
-                "config": op_config,
-            },
-        }
-    }
-
-    @dagster.job(
-        name=f"{name}_health_check_job",
-        description=f"Health check: {kind}",
-        executor_def=dagster.multiprocess_executor.configured({"max_concurrent": policy.max_concurrent}),
-        tags={"owner": owner.value},
-        config=job_config,
-    )
-    def health_check_job():
-        team_batches = get_all_team_ids_op()
-        results = team_batches.map(batch_op)
-        aggregate_op(results.collect())
-
-    schedule_def = None
-    if schedule is not None:
-
-        @dagster.schedule(
-            cron_schedule=schedule,
-            job=health_check_job,
-            execution_timezone="UTC",
-            name=f"{name}_health_check_schedule",
-            tags={"owner": owner.value},
-        )
-        def health_check_schedule(context: dagster.ScheduleEvaluationContext):
-            skip_reason = check_for_concurrent_runs(context, tags={})
-            if skip_reason:
-                return skip_reason
-            return dagster.RunRequest(run_config=job_config)
-
-        schedule_def = health_check_schedule
-
-    return HealthCheckDefinition(job=health_check_job, schedule=schedule_def)
