+++ posthog/temporal/health_checks/activities.py
@@ -11,11 +11,11 @@ import temporalio.activity
 
 from posthog.clickhouse.client.execute import KillSwitchLevel, get_kill_switch_level
 from posthog.clickhouse.query_tagging import Feature, tag_queries
-from posthog.dags.common.health.observability import push_health_check_metrics
 from posthog.models.organization import OrganizationMembership
 from posthog.sync import database_sync_to_async
 from posthog.temporal.common.rollout import filter_ids_for_rollout
 from posthog.temporal.health_checks.models import BatchResult, HealthCheckWorkflowInputs
+from posthog.temporal.health_checks.observability import push_health_check_metrics
 from posthog.temporal.health_checks.processing import _process_batch_detection
 from posthog.temporal.health_checks.registry import ensure_registry_loaded, get_detect_fn, get_product
